requirements.txt
Django==2.2.12
psycopg2-binary==2.8.5
gunicorn==20.1.0
simple-settings==0.19.1
python-dotenv==0.13.0